qemu v8 pr: common: change build targets
Only build PR PK normal world libraries and tests that do not call TEE
functions directly as they have moved (or will be moved eventually) to
the TA and there's no way for these libraries and tests to link to them.

Previously, these libraries and tests were linking to the Pseudo TEE
which is available for linking in normal world and hence caused no build
errors.

Signed-off-by: Victor Chong <victor.chong@linaro.org>