qemu_v8: Add option to try measured boot flow with sw-tpm
Option SWTPM has been added for this flow. It is enabled by default.

rootfs is pre-built for this flow as we want to use the latest
tpm2-tools not available with buildroot.

linux is available in this rootfs, so it has not been built separately.

On u-boot, using boot command would use bootefi and also invoke softtpm.

At linux, do the following to see evenlog
> sh log

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>